+++ src/application/mappers/PartnerMapper.ts
@@ -4,6 +4,7 @@ import { CoverageArea } from '../../domain/models/partner/CoverageArea';
 import { GeojsonType } from '../../domain/models/partner/geojsonType';
 import { Partner } from '../../domain/models/partner/partner';
 import { PartnerApplicationModel } from './PartnerApplicationModel';
+import { PartnerViewModel } from './PartnerViewModel';
 
 export class PartnerMapper {
   public static toDomain(requestPartner: PartnerApplicationModel): Partner {
@@ -31,5 +32,22 @@ export class PartnerMapper {
 
       throw new IllegalArgumentException('Invalid request body: error when try to create model partner');
     }
+  }
+
+  public static toModelView(partner: Partner): PartnerViewModel {
+    return {
+      id: partner.getId(),
+      tradingName: partner.getTradingName(),
+      ownerName: partner.getOwnerName(),
+      document: partner.getDocumentNumber(),
+      coverageArea: {
+        type: partner.getCoverageAreaType(),
+        coordinates: partner.getCoverageCoordinates(),
+      },
+      address: {
+        type: partner.getAddressType(),
+        coordinates: partner.getAddressCoordinates(),
+      },
+    };
   }
 }