+++ src/domain/exceptions/NotFoundPartnerException.ts
@@ -1,0 +1,5 @@
+export class NotFoundPartnerException extends Error {
+  constructor(message: string) {
+    super(message);
+  }
+}