+++ src/infrastructure/utils/Request.ts
@@ -1,6 +1,4 @@
 export class Request {
-  constructor() {}
-
   public static parse(result: any): any {
     return JSON.parse(result);
   }