+++ src/presentation/controllers/PartnerController.ts
@@ -1,6 +1,6 @@
 import { Request } from '../../infrastructure/utils/Request';
 import { PartnerService } from '../../application/services/partner';
-import { PartnerApplicationModel } from '../../application/mappers/PartnerApplicationModel';
+import { PartnerApplicationModel, PartnerApplicationModelId } from '../../application/mappers/PartnerApplicationModel';
 import { Response, ResponseData } from '../../infrastructure/utils/Response';
 
 export class PartnerController {
@@ -10,12 +10,23 @@ export class PartnerController {
     this.partnerService = partnerService;
   }
 
-  public async save(event: any): Promise<ResponseData> {
+  public async post(event: any): Promise<ResponseData> {
     try {
       const partnerApplicationModel = <PartnerApplicationModel>Request.parse(event.body);
-      await this.partnerService.save(partnerApplicationModel);
+      const partnerModelView = await this.partnerService.save(partnerApplicationModel);
 
-      return Response.success({}, 201);
+      return Response.success(partnerModelView, 201);
+    } catch (error) {
+      return Response.error(error);
+    }
+  }
+
+  public async get(event: any): Promise<ResponseData> {
+    try {
+      const partnerApplicationModelId = <PartnerApplicationModelId>event.pathParameters;
+      const partnerModelView = await this.partnerService.find(partnerApplicationModelId);
+
+      return Response.success(partnerModelView, 200);
     } catch (error) {
       return Response.error(error);
     }