+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## 1.2.0 (2021-01-16)
+
+
+### Features
+
+* find partner method ([6299aed](https://github.com/ricardo-lino/test-ze-delivery/commit/6299aed714bae2bb5c8e85ed255d9e9d8287e285))
+* save partner method ([892964c](https://github.com/ricardo-lino/test-ze-delivery/commit/892964cdf1ffed3eb67615936030e6de8321c686))
+
 ## 1.1.0 (2021-01-16)